Adding an intro screen
When starting the app for the first time, show an intro screen
describing how the app works.
