+++ pubspec.yaml
@@ -16,7 +16,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.0.1+6
+version: 1.0.2+7
 
 environment:
   sdk: '>=3.0.0'
@@ -46,6 +46,8 @@ dependencies:
   dio: ^5.3.2
 
   image: ^4.0.15
+  shared_preferences: ^2.2.1
+  intro_slider: ^4.2.1
 
 dev_dependencies:
   flutter_test:
@@ -83,6 +85,10 @@ flutter:
     - assets/empty.png
     - assets/pin.png
     - assets/binoculars.png
+    - assets/intro/mont_blanc_1.png
+    - assets/intro/mont_blanc_2.png
+    - assets/intro/mont_blanc_3.png
+    - assets/intro/loading_map.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
